+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This changelog covers all three packages, as they are (for now) updated as a who
 ## Unreleased
 
 - Make search result previews smaller
+- Fetch full collections when showing CollectionCard
+- `useResource` defaults to not accepting incomplete resources
 
 ## v0.30.0